[BUG-3] Rename namespace of module :common:resources
